pyproject.toml
[tool.poetry]
name = "ibpy"
version = "0.3.0"
description = ""
authors = ["Jacek Palczewski <jpalczewski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytz = "^2019.3"
Django = "^3.0.3"
django-configurations = "^2.2"
gunicorn = "^20.0.4"
newrelic = "^5.4.1.134"
psycopg2-binary = "^2.8.4"
dj-database-url = "^0.5.0"
django-model-utils = "^4.0.0"
django_unique_upload = "^0.2.1"
djangorestframework = "^3.11.0"
Markdown = "^3.1.1"
django-filter = "^2.2.0"
ipdb = "^0.12.3"
ipython = "^7.11.1"
mkdocs = "^1.0.4"
flake8 = "^3.7.9"
mock = "^3.0.5"
factory-boy = "^2.12.0"
django-nose = "^1.4.6"
nose-progressive = "^1.5.2"
coverage = "^5.0.3"
django-storages = "^1.8.0"
boto3 = "^1.11.1"
python-semantic-release = "^6.0.1"
whitenoise = "^5.0.1"
django-rest-framework-social-oauth2 = "^1.1.0"
python-dotenv = "^0.13.0"
django-cors-headers = "^3.3.0"
pyyaml = "^5.3.1"
uritemplate = "^3.0.1"

[tool.poetry.dev-dependencies]

[tool.semantic_release]
version_variable="pyproject.toml:version"
hvcs="github"
upload_to_release=true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
